Add publishing to Moteus wrapper
