+++ ROS/controls/msg/MotorOrientation.msg
@@ -1,3 +1,2 @@
 int32 numMotors
-int32 numAttributes
-float64[] attributes
+MotorState[] states
