+++ ROS/controls/msg/MotorState.msg
@@ -1,0 +1,3 @@
+float64 position
+float64 velocity
+float64 torque
