+++ ROS/controls/src/raspberry_controls.py
@@ -1,9 +1,11 @@
+#!/usr/bin/env python3
 from typing import List
 from controls.msg import MotorOrientation
-from motor_descriptor import MoteusMotorDescriptor
+from controls.msg import MotorState
 import rospy
-import moteus
 import controls_constants as constants
+import sys
+from Moteus import Moteus
 
 class RaspberryPiMoteusWrapper:
     """
@@ -32,19 +34,15 @@ class RaspberryPiMoteusWrapper:
     Attributes
     ----------
     hw_refresh_rate : float
-        The rate, in seconds, at which the moteus hardware is queried
-        for state changes
+        The rate, in Hertz, at which readings of the Moteus
+        motors for their current respective statuses are taken
 
     num_motors : int
-        The number of motors that are controlled
+        The number of motors that are controller
 
-     name_map : { str : MoteusMotorDescriptor }
-        Maps the names given to each motor at initialization-time to a
-        descriptor describing how that motor is configured
-
-    id_map : { int : MoteusMotorDescriptor }
-        Maps the hardware ids given to each motor to a
-        descriptor describing how that motor is configured
+    id_map : { int : int }
+        Maps the hardware ids given to each motor to the
+        CAN bus lane which that motor communicates on
 
     -------
     Methods
@@ -56,43 +54,40 @@ class RaspberryPiMoteusWrapper:
 
     # MARK: - Constructors -
 
-    def __init__(self, motor_list: List[MoteusMotorDescriptor], hw_refresh_rate: float = constants.CONTROLS_PKG_DEFAULT_HW_REFRESH_RATE):
+    def __init__(self, motor_list, hw_refresh_rate: float = constants.CONTROLS_PKG_DEFAULT_HW_REFRESH_RATE):
         """Construct a new wrapper around the Moteus class
 
-        @param motor_list: A list of descriptors describing how each motor is arranged
+        @param motor_list: A list which maps each motor id to its
+        respective CAN bus lane
         
-        @param hw_refresh_rate: The rate, in HZ, at which the Moteus motors
-        is pinged for their current respective statuses
+        @param hw_refresh_rate: The rate, in Hertz, at which readings of the Moteus
+        motors for their current respective statuses are taken
         """
         if len(motor_list) == 0:
             raise ValueError('The wrapper class expects at least one motor')
 
         self.hw_refresh_rate = hw_refresh_rate
         self.num_motors = len(motor_list)
-        self.name_map = {}
-        self.id_map = {}
+        self.id_map = motor_list
 
         # The arrangement of motor ids into CAN lanes can be computed
-        # by processing the descriptors one-by-one
+        # by processing the mapping of motors IDs to the CAN bus lanes
         moteus_ids = []
         current_max_lane_id = -1
 
-        # We build all of data from the descriptors at once
+        # We build all of data from the mapping at once
         # instead of using list/dictionary comprehensions
         # to reduce the number of iterations through the list
-        for motor_desc in motor_list:
-            self.name_map[motor_desc.name] = motor_desc
-            self.id_map[motor_desc.hardware_id] = motor_desc
-            lane_id = motor_desc.can_bus_lane_id
-
+        for motor_id, can_bus_lane_id in motor_list.items():
             # Determines if we need to add more lanes
-            if lane_id > current_max_lane_id:
-                # Add (lane_id - current_max_lane_id) more lanes
-                moteus_ids += [[] for _ in range(current_max_lane_id, lane_id)]
+            if can_bus_lane_id > current_max_lane_id:
+                # Add (can_bus_lane_id - current_max_lane_id) more lanes
+                moteus_ids += [[] for _ in range(current_max_lane_id, can_bus_lane_id)]
 
-            moteus_ids[lane_id] = motor_desc.hardware_id
+            moteus_ids[can_bus_lane_id].append(motor_id)
 
-        self.moteus_instance = moteus.Moteus(moteus_ids)
+        # Note that the call to the initializer will block
+        self.moteus_instance = mot.Moteus(moteus_ids, simulation=True)
 
         self.desired_orientation_sub = rospy.Subscriber(
             constants.CONTROLS_PKG_DESIRED_ORIENTATION_TOPIC,
@@ -109,30 +104,16 @@ class RaspberryPiMoteusWrapper:
 
     # MARK: - Methods -
 
-    def motor_description_for_name(self, name: str):
-        """Fetches the motor descriptor corresponding to the
-        tag _name_ given to some motor at initialization time
-
-        @param name: The name identifying the motor whose
-        descriptor should be fetched
-
-        @return: A MoteusMotorDescriptor which describes how the
-        motor named _name_ communicates with this process
-        """
-        return self.name_map[name]
-
     def motor_description_for_id(self, hwid: int):
-        """Fetches the motor descriptor corresponding to the
+        """Fetches the CAN bus lane ID corresponding to the
         motor with ID _id_
 
         @param hwid: The hardware ID identifying the motor whose
         descriptor should be fetched
 
-        @return: A MoteusMotorDescriptor which describes how the
-        motor with ID _id_ communicates with this process
+        @return: An identifier naming a distinct CAN bus lane
         """
         return self.id_map[hwid]
-
 
     # MARK: - ROS Messages -
 
@@ -147,10 +128,20 @@ class RaspberryPiMoteusWrapper:
         self.moteus_instance.waitUntilReady()
 
         while (not rospy.is_shutdown()):
-            latest_moteus_hardware_state = self.moteus_instance.getParsedResults()
+            latest_motor_state = self.moteus_instance.getParsedResults()
+            motor_states = []
 
-            # Publish here
+            # Each 'motor_desc' is a dictionary corresponding
+            # to the state of a motor
+            for motor_desc in latest_motor_state:
+                motor_state = MotorState()
+                motor_state.position = motor_desc['POSITION']
+                motor_state.velocity = motor_desc['VELOCITY']
+                motor_state.torque = motor_desc['TORQUE']
+                motor_states.append(motor_state)
 
+            msg = MotorOrientation(numMotors=self.num_motors, states=motor_states)
+            self.current_orientation_publisher.publish(msg)
             rate.sleep()
 
     def __desired_orientation_callback(self, msg: MotorOrientation):
@@ -162,28 +153,28 @@ class RaspberryPiMoteusWrapper:
         of type MotorOrientation describing how the motors
         should be re-oriented
         """
-        for motor_id in range(msg.numMotors):
-            pos = msg.attributes[3 * motor_id]
-            vel = msg.attributes[3 * motor_id + 1]
-            torque = msg.attributes[3 * motor_id + 2]
-            self.moteus_instance.setAttributes(motor_id, pos=pos, velocity=vel, torque=torque)
+        if len(id_map) != msg.num_motors:
+            raise ValueError('Expected a exactly one desired orientation for each motor')
+
+        # The state of the motor at index 'msg_index' corresponds to the
+        # (msg_index + 1)st motor in the dictionary mapping motor ids to
+        # their CAN bus lanes
+        for msg_index, motor_id in enumerate(id_map.keys()):
+            motor_state = msg.states[motor_id]
+            self.moteus_instance.setAttributes(motor_id, pos=motor_state.position, velocity=motor_state.velocity, torque=motor_state.torque)
 
     # MARK - Python Magic Methods -
 
     def __str__(self):
         """Outputs a formatted string describing the instance"""
         name = 'RaspberryPiMoteusWrapper\n'
-        subscribed_to = 'subscribed to: ' + RaspberryPiMoteusWrapper.SUBSCR_TOPIC_NAME + "\n"
-        pub_to = 'publishing to: ' + RaspberryPiMoteusWrapper.PUB_TOPIC_NAME
+        subscribed_to = 'subscribed to: ' + constants.CONTROLS_PKG_DESIRED_ORIENTATION_TOPIC + "\n"
+        pub_to = 'publishing to: ' + constants.CONTROLS_PKG_CURRENT_ORIENTATION_TOPIC
         return name + subscribed_to + pub_to
 
-if __name__ == 'main':
-    motor_left = MoteusMotorDescriptor('left-motor', 1, 0)
-    motor_middle = MoteusMotorDescriptor('middle-motor', 2, 1)
-    motor_right = MoteusMotorDescriptor('right-motor', 3, 1)
+# Connect to the ROS subsystem
+rospy.init_node(constants.CONTROLS_PKG_RASPBERRYPI_NODE_NAME)
 
-    wrapper = RaspberryPiMoteusWrapper(
-        motor_list=[motor_left, motor_middle, motor_right], 
-        hw_refresh_rate=0.1
-    )
-    wrapper.run_motor_orientation_loop()
+# Run the node
+wrapper = RaspberryPiMoteusWrapper(motor_list=constants.CONTROLS_PKG_BUS_CONFIGURATION)
+wrapper.run_motor_orientation_loop()
